+++ web/config/index.ts
@@ -182,7 +182,6 @@ export const CSRF_HEADER_NAME = 'X-CSRF-Token'
 export const ACCESS_TOKEN_LOCAL_STORAGE_NAME = 'access_token'
 export const PASSPORT_LOCAL_STORAGE_NAME = (appCode: string) => `passport-${appCode}`
 export const PASSPORT_HEADER_NAME = 'X-App-Passport'
-export const ACEDATACLOUD_OAUTH_SESSION_LOCAL_STORAGE_NAME = 'acedatacloud_oauth_session'
 
 export const WEB_APP_SHARE_CODE_HEADER_NAME = 'X-App-Code'
